* Switch hashing method * Add flags to clear users/cache in server_config.py

+++ server/DatabaseHandler.py
@@ -2,7 +2,8 @@ from pymongo.mongo_client import MongoClient
 from pymongo.server_api import ServerApi
 import time
 import math
-from server_config import databaseUri
+from hashlib import sha256
+from server_config import databaseUri, clear_users_on_start, clear_cache_on_start
 
 
 class DatabaseHandler:
@@ -34,9 +35,9 @@ class DatabaseHandler:
         if 'users' in self.userDb.list_collection_names():
             self.userCollection = self.userDb.get_collection('users')
 
-            # TODO: Temporary (wipe the db on startup):
-            self.userCollection.drop()
-            self.initUsersCollection()
+            if clear_users_on_start:
+                self.userCollection.drop()
+                self.initUsersCollection()
         else:
             self.userCollection = self.userDb.create_collection('users')
 
@@ -47,9 +48,9 @@ class DatabaseHandler:
         if 'cache' in self.cacheDb.list_collection_names():
             self.cacheCollection = self.cacheDb.get_collection('cache')
 
-            # TODO: Temporary (wipe the db on startup):
-            self.cacheCollection.drop()
-            self.initCacheCollection()
+            if clear_cache_on_start:
+                self.cacheCollection.drop()
+                self.initCacheCollection()
         else:
             self.cacheCollection = self.cacheDb.create_collection('cache')
 
@@ -69,7 +70,7 @@ class DatabaseHandler:
     ### CACHE ###
 
     def findCachedSummary(self, long_description):
-        descriptionHash = hash(long_description)
+        descriptionHash = sha256(long_description.encode("utf-8")).hexdigest()
         filter = {"description": descriptionHash}
         item = self.cacheCollection.find_one(filter)
         if item and 'summary' in item: 
@@ -78,7 +79,7 @@ class DatabaseHandler:
             return None
         
     def saveCachedSummary(self, long_description, summary):
-        descriptionHash = hash(long_description)
+        descriptionHash = sha256(long_description.encode("utf-8")).hexdigest()
         item = {"description": descriptionHash, "summary": summary}
         self.cacheCollection.insert_one(item)
